foldercomumnomecorreto/requirements.txt

# Function dependencies, for example:
# package>=version
google-cloud-pubsub
google-cloud-storage